[Frontend] create new authentification page

+++ Frontend/src/store/modules/users.js
@@ -51,7 +51,7 @@ export default {
     },
     async createUser({commit}, payload){
       axios.post("/users", payload).then((response) => {
-        if (response.data){
+        if (response && response.data){
           localStorage.clear();
           commit("auth", response);
           router.push("/");
